+++ lib/d_input.dart
@@ -2,8 +2,8 @@ library d_input;
 
 import 'package:flutter/material.dart';
 
+part 'icon_spec.dart';
 part 'input.dart';
+part 'input_dropdown.dart';
 part 'input_mix.dart';
 part 'input_password.dart';
-part 'icon_spec.dart';
-part 'input_dropdown.dart';