+++ lib/input_dropdown.dart
@@ -1,14 +1,27 @@
 part of 'd_input.dart';
 
-class DInputDropdown<T> extends StatelessWidget {
+class DInputDropdown<T> extends StatefulWidget {
   const DInputDropdown({
     super.key,
     required this.value,
     required this.items,
     required this.inputOnChanged,
     this.boxRadius = 20,
-    this.boxBorder,
+    this.focusedBoxRadius = 20,
+    this.boxBorder = const Border.fromBorderSide(
+      BorderSide(
+        color: Colors.grey,
+        width: 1,
+      ),
+    ),
+    this.focusedBoxBorder = const Border.fromBorderSide(
+      BorderSide(
+        color: Colors.grey,
+        width: 2,
+      ),
+    ),
     this.boxColor,
+    this.focusedBoxColor,
     this.noBoxBorder = false,
     this.inputPadding = const EdgeInsets.symmetric(
       horizontal: 20,
@@ -46,21 +59,45 @@ class DInputDropdown<T> extends StatelessWidget {
   /// default: 20
   final double boxRadius;
 
+  /// radius for all corner (box wrapper)
+  ///
+  /// default: 20
+  final double focusedBoxRadius;
+
   /// styling for box border
   ///
   /// default:
   /// ```dart
-  /// Border.all(
-  ///   color: Theme.of(context).primaryColor,
-  ///   width: 2,
+  /// const Border.fromBorderSide(
+  ///   BorderSide(
+  ///     color: Colors.grey,
+  ///     width: 1,
+  ///   ),
   /// )
   /// ```
-  final BoxBorder? boxBorder;
+  final BoxBorder boxBorder;
+
+  /// styling for box border
+  ///
+  /// ```dart
+  /// const Border.fromBorderSide(
+  ///   BorderSide(
+  ///     color: Colors.grey,
+  ///     width: 2,
+  ///   ),
+  /// )
+  /// ```
+  final BoxBorder focusedBoxBorder;
 
   /// background color
   ///
   /// default: Theme.of(context).colorScheme.surfaceContainer
   final Color? boxColor;
+
+  /// background color
+  ///
+  /// default: Theme.of(context).colorScheme.surfaceContainer
+  final Color? focusedBoxColor;
 
   /// default: false
   final bool noBoxBorder;
@@ -132,71 +169,101 @@ class DInputDropdown<T> extends StatelessWidget {
   final List<Widget>? rightChildren;
 
   @override
+  State<DInputDropdown<T>> createState() => _DInputDropdownState<T>();
+}
+
+class _DInputDropdownState<T> extends State<DInputDropdown<T>> {
+  final listenFocus = ValueNotifier(false);
+
+  @override
+  void dispose() {
+    listenFocus.dispose();
+    super.dispose();
+  }
+
+  @override
   Widget build(BuildContext context) {
     final inputBorder = OutlineInputBorder(
-      borderRadius: BorderRadius.circular(inputRadius),
-      borderSide: inputBorderSide,
+      borderRadius: BorderRadius.circular(widget.inputRadius),
+      borderSide: widget.inputBorderSide,
     );
-    final newBoxBorder = boxBorder ??
-        Border.all(
-          color: Theme.of(context).primaryColor,
-          width: 2,
-        );
+    final defaultBoxColor = Theme.of(context).colorScheme.surfaceContainer;
+
     return Column(
-      crossAxisAlignment: crossAxisAlignmentTitle,
+      crossAxisAlignment: widget.crossAxisAlignmentTitle,
       children: [
-        if (title != null)
+        if (widget.title != null)
           Padding(
-            padding: EdgeInsets.only(bottom: titleGap),
+            padding: EdgeInsets.only(bottom: widget.titleGap),
             child: Text(
-              title!,
-              style: titleStyle,
+              widget.title!,
+              style: widget.titleStyle,
             ),
           ),
-        DecoratedBox(
-          decoration: BoxDecoration(
-            color: boxColor ?? Theme.of(context).colorScheme.surfaceContainer,
-            borderRadius: BorderRadius.circular(boxRadius),
-            border: noBoxBorder ? null : newBoxBorder,
-          ),
-          child: Row(
-            crossAxisAlignment: crossAxisAlignmentBox,
-            children: [
-              prefixIcon.build(context, boxRadius),
-              if (leftChildren != null) ...leftChildren!,
-              Expanded(
-                child: Padding(
-                  padding: inputMargin,
-                  child: DropdownButtonFormField<T>(
-                    value: value,
-                    items: items,
-                    style: inputStyle,
-                    onChanged: inputOnChanged,
-                    borderRadius: BorderRadius.circular(boxRadius),
-                    itemHeight: itemHeight,
-                    menuMaxHeight: menuMaxHeight,
-                    icon: icon,
-                    dropdownColor: dropdownColor ??
-                        Theme.of(context).colorScheme.surfaceContainer,
-                    decoration: InputDecoration(
-                      filled: inputBackgroundColor != null,
-                      fillColor: inputBackgroundColor,
-                      isDense: true,
-                      contentPadding: inputPadding,
-                      border: inputBorder,
-                      errorBorder: inputBorder,
-                      enabledBorder: inputBorder,
-                      focusedBorder: inputBorder,
-                      disabledBorder: inputBorder,
-                      focusedErrorBorder: inputBorder,
+        ListenableBuilder(
+          listenable: listenFocus,
+          builder: (context, child) {
+            bool isFocus = listenFocus.value;
+            return DecoratedBox(
+              decoration: BoxDecoration(
+                color: isFocus
+                    ? widget.focusedBoxColor ??
+                        (widget.boxColor ?? defaultBoxColor)
+                    : widget.boxColor ?? defaultBoxColor,
+                borderRadius: BorderRadius.circular(
+                  isFocus ? widget.focusedBoxRadius : widget.boxRadius,
+                ),
+                border: widget.noBoxBorder
+                    ? null
+                    : isFocus
+                        ? widget.focusedBoxBorder
+                        : widget.boxBorder,
+              ),
+              child: Row(
+                crossAxisAlignment: widget.crossAxisAlignmentBox,
+                children: [
+                  widget.prefixIcon.build(context, widget.boxRadius),
+                  if (widget.leftChildren != null) ...widget.leftChildren!,
+                  Expanded(
+                    child: Padding(
+                      padding: widget.inputMargin,
+                      child: Focus(
+                        onFocusChange: (value) {
+                          listenFocus.value = value;
+                        },
+                        child: DropdownButtonFormField<T>(
+                          value: widget.value,
+                          items: widget.items,
+                          style: widget.inputStyle,
+                          onChanged: widget.inputOnChanged,
+                          borderRadius: BorderRadius.circular(widget.boxRadius),
+                          itemHeight: widget.itemHeight,
+                          menuMaxHeight: widget.menuMaxHeight,
+                          icon: widget.icon,
+                          dropdownColor: widget.dropdownColor ??
+                              Theme.of(context).colorScheme.surfaceContainer,
+                          decoration: InputDecoration(
+                            filled: widget.inputBackgroundColor != null,
+                            fillColor: widget.inputBackgroundColor,
+                            isDense: true,
+                            contentPadding: widget.inputPadding,
+                            border: inputBorder,
+                            errorBorder: inputBorder,
+                            enabledBorder: inputBorder,
+                            focusedBorder: inputBorder,
+                            disabledBorder: inputBorder,
+                            focusedErrorBorder: inputBorder,
+                          ),
+                        ),
+                      ),
                     ),
                   ),
-                ),
+                  if (widget.rightChildren != null) ...widget.rightChildren!,
+                  widget.suffixIcon.build(context, widget.boxRadius),
+                ],
               ),
-              if (rightChildren != null) ...rightChildren!,
-              suffixIcon.build(context, boxRadius),
-            ],
-          ),
+            );
+          },
         ),
       ],
     );